requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
supabase==2.8.0
python-dotenv==1.0.0
Pillow==10.1.0
gunicorn==21.2.0
Werkzeug==3.0.1
